third_party/WebKit/LayoutTests/fast/animation/request-animation-frame-cancel-expected.txt

This tests cancelling a requestAnimationFrame callback

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS successfullyParsed is true

TEST COMPLETE

PASS callbackFired is false
PASS successfullyParsed is true

TEST COMPLETE
